requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
Flask==3.1.3
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
Mako==1.3.11
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
